Don't use MultiinterfaceController
This is no longer necessary, because the TrajectoryInterface provides
read/write access to the trajectory buffers.
I previously thought that MultiinterfaceController controllers were
necessary to also get access to the individual joint states.

Our approach instead is to use the feedback mechanism of the
TrajectoryInterface.
Checking for path and goal tolerances is done solely with this feedback,
whose type offers all relevant fields for error checking.

This should work both for joint-based and Cartesian
PassThroughControllers.