Implement tolerance checking and goal finishing
The feedback's error is used for path and goal tolerance checking.

The action finishes once the last trajectory point's time_from_start is
reached (plus possible goal_time_tolerance). That's not absolutely
accurate, though, because the vendor trajectory controller might take
some time to receive and process the forwarded goal.

But I guess that's as accurate as it gets with PassThroughControllers..