Add a control policy for Cartesian trajectory publishing
This adds an additional `CartesianTrajectoryPublisher` as read-only
controller.  It uses and offers the CartesianTrajectoryController's
action interface but does not claim resources nor actuates the robot's
joints.  It just publishes the trajectory's interpolated Cartesian state
as reference PoseStamped and TwistStamped messages.  Its purpose is to
offer an easy way for users to test whether their Cartesian trajectories
_look_ as expected.