Make robot communication closed-loop
This introduces a design change for how the PassThroughControllers
interact with the robot. They reject invalid trajectories like before.
But when canceling running trajectory execution, e.g. by user preemption request or
failing path tolerances, they ultimately wait for the
robot to confirm this with using the done callback mechanism.

This makes sure that the PassThroughControllers do not signal successful preemption
to action clients while the robot is still moving or behaving unresponsive.

The centralized handling of the _done_ signal by the robot loses some
insights on the actual error.
This is compensated with ROS_WARN notifications where the error appears
in the PassThroughControllers.